kano-init: Fixing GUI logout problem
After adding the exit command, the user was still put into root shell
after logging out of LXDE. The problem is that the script is run in a
subshell, not the login shell itself.

I experimented with a few approaches. When kano-init is used directly,
without a shell in /etc/inittab, there are many problems with commands,
because without a login shell, the ENV vars don't get initialised
properly.

Ultimately, I decided to use the login shell as root to get the benefits
and kill it using a trick from the autostart-kano-init.sh profile script.