kano-init: Adding a check for existing username
If the user tries to enter an existing username (such as root), the
script would die with error.

This commit should fix that problem. It also counts with the case when
the subsequent user picks the same name after reset as did the one
before (this case is allowed).

Issue #126 (https://github.com/KanoComputing/Spindle/issues/126)