Project i18n and Argentinian Spanish translations
The changes here were two fold. Firstly, all the strings in the
project are marked for i18n, subprocesses are launched with their
respective translations, and binaries are now installing the
translation package. Secondly, the Makefile for po files has been
updated to extract from Python, similarly to all of our other
projects. The en_QQ was also generated to test out the app live.